+++ bin/main/assets/qtr4tdm/models/item/base_expander0.json
@@ -1,0 +1,18 @@
+{
+    "parent": "builtin/generated",
+    "textures": {
+        "layer0": "qtr4tdm:items/baseexpandplate1"
+    },
+    "display": {
+        "thirdperson": {
+            "rotation": [ -90, 0, 0],
+            "translation": [ 0, 1, -3 ],
+            "scale": [ 0.25, 0.25, 0.25]
+        },
+        "firstperson": {
+            "rotation": [ 0, -135, 25],
+            "translation": [ 0, 4, 2],
+            "scale": [ 1, 1, 1]
+        }
+    }
+}
