+++ bin/main/assets/qtr4tdm/models/item/base_expander3.json
@@ -1,18 +1,18 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "qtr4:items/baseexpandplate1"
+        "layer0": "qtr4tdm:items/baseexpandplate4"
     },
     "display": {
         "thirdperson": {
             "rotation": [ -90, 0, 0],
             "translation": [ 0, 1, -3 ],
-            "scale": [ 0.55, 0.55, 0.55]
+            "scale": [ 0.25, 0.25, 0.25]
         },
         "firstperson": {
             "rotation": [ 0, -135, 25],
             "translation": [ 0, 4, 2],
-            "scale": [ 1.7, 1.7, 1.7]
+            "scale": [ 1, 1, 1]
         }
     }
 }
