+++ src/main/java/com/anet/qtr4tdm/common/items/BaseExpandItem.java
@@ -14,21 +14,22 @@ public class BaseExpandItem extends Item implements IMetadataItem {
         this.setHasSubtypes(true);
         this.setRegistryName("base_expander");
         this.setMaxDamage(0);
-        this.setCreativeTab(TdmMod.qtr4);
+        //this.setCreativeTab(TdmMod.qtr4);
     }
-
+    
     @Override
     public String getUnlocalizedName(ItemStack item) {
         int metadata = item.getItemDamage();
-        return "base_expander_lvl" + metadata;
+        return "base_expander_lvl" + (metadata + 1);
     }
 
     @Override
     public void getSubItems(CreativeTabs tab, NonNullList<ItemStack> items) {
-        for (int level = 0; level < 3; level++) {
+        super.getSubItems(tab, items);
+        if (tab != TdmMod.qtr4) return;
+        for (int level = 0; level < GetMetaCount(); level++) {
             items.add(new ItemStack(BlocksInit.BASEEXPANDER, 1, level));
         }
-        super.getSubItems(tab, items);
     }
 
     @Override
