+++ src/main/java/com/anet/qtr4tdm/uebki/gui/EnergyConsumerGui.java
@@ -6,7 +6,6 @@ import com.anet.qtr4tdm.uebki.messages.BasedRequest;
 
 import org.lwjgl.opengl.GL11;
 
-import net.minecraft.client.gui.FontRenderer;
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.util.ResourceLocation;
