+++ src/main/resources/assets/qtr4tdm/textures/items/baseexpandplate3_e.png
