+++ src/main/resources/assets/qtr4tdm/textures/items/radar_base.png
